requirements.txt
pytorchvideo==0.1.5
pytorch-lightning==1.9.4
editdistance==0.6.0
scikit-learn==0.24.2
psutil==5.9.0
opencv-python==4.5.3.56
einops==0.3.0
decord==0.6.0
lmdb==1.2.1
imutils==0.5.4
submitit==1.3.3
pandas==1.3.5
dotmap
